Switch names of Domain Attribute screen. Marked old screen as obsolete.
